More windows friendly.
Make sys_getloadavg adjustments as windows php instances do not have
this available. Make sure the LC_MESSAGES is being specified to be an
integer. Add fog link to ignored items.

+++ packages/web/lib/pages/processlogin.class.php
@@ -168,7 +168,7 @@ class ProcessLogin extends FOGPage
         );
         $this->_specLang();
         setlocale(
-            LC_MESSAGES,
+            (int)LC_MESSAGES,
             sprintf(
                 '%s.UTF-8',
                 self::$locale
